backend/requirements.txt

flask
flask-cors
web3
python-docx
pypdf2
